プロフィールの編集を完了した際に即座に変更内容を適応させる
